add new deltastream class
